drone_manual_control/requirements.txt
numpy>=1.21.0
pyyaml>=6.0
dataclasses>=0.6; python_version < "3.7"
typing-extensions>=4.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.0.0 